refactor: move sourcemap to it's own file
