+++ projects/openapi-io/src/index.ts
@@ -1,13 +1,16 @@
 import { inGit, loadSpecFromBranch } from './loaders/file-on-branch';
 import { loadSpecFromFile, loadSpecFromUrl } from './loaders/file';
 import {
-  JsonPath,
-  JsonSchemaSourcemap,
   JSONParserError,
   parseOpenAPIFromRepoWithSourcemap,
   ParseOpenAPIResult,
   parseOpenAPIWithSourcemap,
 } from './parser/openapi-sourcemap-parser';
+import {
+  JsonPath,
+  JsonSchemaSourcemap,
+  resolveJsonPointerInYamlAst,
+} from './parser/sourcemap';
 import { loadYaml, isYaml, isJson, writeYaml } from './write/index';
 import {
   watchDependencies,
@@ -22,6 +25,7 @@ export {
   loadSpecFromBranch,
   parseOpenAPIWithSourcemap,
   parseOpenAPIFromRepoWithSourcemap,
+  resolveJsonPointerInYamlAst,
   JsonSchemaSourcemap,
   JSONParserError,
   loadYaml,
