+++ projects/openapi-utilities/src/render/ast-helpers.ts
@@ -1,40 +1,6 @@
 import {
-  Kind,
-  YamlMap,
-  YAMLMapping,
   YAMLNode,
-  YAMLSequence,
 } from 'yaml-ast-parser';
-import { jsonPointerHelpers } from '@useoptic/json-pointer-helpers';
-
-export function resolveJsonPointerInYamlAst(
-  node: YAMLNode, // root ast
-  pointer: string
-): YAMLNode | undefined {
-  const decoded = jsonPointerHelpers.decode(pointer);
-  const isEmpty =
-    decoded.length === 0 || (decoded.length === 1 && decoded[0] === '');
-
-  if (isEmpty) return node;
-
-  const found: YAMLNode | undefined = decoded.reduce((current, path) => {
-    if (!current) return undefined;
-    const node: YAMLNode = current.key ? current.value : current;
-    const isNumericalKey =
-      !isNaN(Number(path)) && (node as any).hasOwnProperty('items');
-
-    if (isNumericalKey) {
-      return (node as YAMLSequence).items[Number(path)];
-    } else {
-      const field = node.mappings.find(
-        (i: YAMLMapping) => i.key.value === path
-      );
-      return field;
-    }
-  }, node as YAMLNode | undefined);
-
-  return found;
-}
 
 export function findLinesForAstAndContents(
   astNode: YAMLNode,
